fix: Adjust padding in AccountControl CSS for consistent layout; add icon generator utility functions
